refactor: Rename html/stylesheet.rs to html/style_sheet.rs
